factored out ResultsFilter.js into separate filters
